Management/requirements.txt
Django==2.1.4
django-rest-multiple-models==2.1.3
djangorestframework==3.9.0
pytz==2018.7
